Fix creation of all chain file
Problem is that it doesn't get created if key file and cert already exists and
are not deployed by nginx state.
So ensure all file always exists.
However I am not sure how this will behave if key and cert will get changed.
Maybe it would be better to use file.managed and template instead.